Add support for storing themes as a constant
